requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.3.3
motor==3.1.1
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentry-sdk==1.32.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1